requirements.freeze.txt
aniso8601==7.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
autoflake==2.2.1
awesome-slugify==1.6.5
black==23.11.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
cryptography==41.0.6
cssselect==1.2.0
cssutils==2.9.0
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-anymail==10.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.4
django-fsm==2.8.1
django-graphql-jwt==0.4.0
django-memoize==2.3.1
django-minio-storage==0.5.7
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
django-storages==1.14.2
django-stubs==4.2.6
django-stubs-ext==4.2.5
django-webpack-loader==2.0.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
factory-boy==3.3.0
Faker==20.1.0
filelock==3.13.1
flake8==6.1.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphene==2.1.9
graphene-django==2.16.0
graphene-validator==0.0.1
graphql-core==2.3.2
graphql-relay==2.0.1
identify==2.5.32
idna==3.6
iniconfig==2.0.0
ipython==8.18.1
isort==6.0.0b2
jedi==0.19.1
lxml==4.9.3
matplotlib-inline==0.1.6
mccabe==0.7.0
minio==7.2.0
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.9.0
phonenumbers==8.13.26
platformdirs==4.0.0
pluggy==1.3.0
pre-commit==3.5.0
premailer==3.10.0
promise==2.3
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pycurl==7.45.2
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
Rx==1.6.3
sentry-sdk==1.37.1
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tomli==2.0.1
traitlets==5.14.0
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.8.0
Unidecode==0.4.21
uritemplate==4.1.1
urllib3==1.26.18
uWSGI==2.0.23
virtualenv==20.24.7
wcwidth==0.2.12
websocket-client==1.6.4
whitenoise==6.6.0
